feat: change hotel capacity type from Integer to Long across relevant classes

+++ src/main/java/com/rzodeczko/infrastructure/persistence/entity/HotelEntity.java
@@ -20,5 +20,5 @@ public class HotelEntity {
     @EqualsAndHashCode.Include
     private Long id;
 
-    private Integer capacity;
+    private Long capacity;
 }
